pyproject.toml
[tool.poetry]
name = "hohohohohoho"
version = "0.1.0"
description = ""
authors = ["frutez <glebka.yatsenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
